refactor: create styles folder
