+++ src/components/Background.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../Background.css';
+import '../styles/Background.css';
 
 export default function Background() {
   return (
