+++ src/components/ProgressBar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../ProgressBar.css';
+import '../styles/ProgressBar.css';
 import currencyFormat from '../helpers/currencyFormat';
 
 export default function ProgressBar({
